+++ CHANGELOG.md
@@ -3,6 +3,7 @@ x.y.z Release notes (yyyy-MM-dd)
 ### Enhancements
 * Improve performance of creating Projection objects and of change
   notifications on projections ([PR #8050](https://github.com/realm/realm-swift/pull/8050)).
+* Allow initialising any sync configuration with `cancelAsyncOpenOnNonFatalErrors`.
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-swift/issues/????), since v?.?.?)